cli: retire retrieve_mpraw_data
